Moved footer into single folder
