+++ src/components/index.jsx
@@ -1,6 +1,6 @@
 export { default as Navbar } from './Navbar';
 export { default as Modal } from './Modal';
-export { default as Footer } from './Footer';
+export { default as Footer } from './Bottem/Footer';
 export { default as Map } from './Map';
 export { default as Contactsbox } from './Contactsbox';
 export { default as PDCalendar } from './PlayDateCalendar/PDCalendar';
